third_party/WebKit/LayoutTests/vibration/vibration-durations-expected.txt

Tests for how durations are handled in the Vibration API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.vibrate([]) is true
PASS areArraysEqual(internals.pendingVibrationPattern(document), []) is true
PASS internals.isVibrating(document) is false
PASS navigator.vibrate(0) is true
PASS areArraysEqual(internals.pendingVibrationPattern(document), []) is true
PASS internals.isVibrating(document) is false
PASS navigator.vibrate([0]) is true
PASS areArraysEqual(internals.pendingVibrationPattern(document), []) is true
PASS internals.isVibrating(document) is false
PASS navigator.vibrate(10000) is true
PASS areArraysEqual(internals.pendingVibrationPattern(document), [10000]) is true
PASS internals.isVibrating(document) is true
PASS navigator.vibrate(10001) is true
PASS areArraysEqual(internals.pendingVibrationPattern(document), [10000]) is true
PASS internals.isVibrating(document) is true
PASS navigator.vibrate([10001, 10001, 10001]) is true
PASS areArraysEqual(internals.pendingVibrationPattern(document), [10000, 10000, 10000]) is true
PASS internals.isVibrating(document) is true
PASS navigator.vibrate(-1) is true
PASS areArraysEqual(internals.pendingVibrationPattern(document), []) is true
PASS internals.isVibrating(document) is false
PASS navigator.vibrate([-1, -1, -1]) is true
PASS areArraysEqual(internals.pendingVibrationPattern(document), [10000, 10000, 10000]) is true
PASS internals.isVibrating(document) is true
PASS successfullyParsed is true

TEST COMPLETE
